change and add pushMethod
