--- pkg/metrics/scrape/scraper.go
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package scrape
-
-import (
-	contextHttp "github.com/apache/dubbo-go-pixiu/pixiu/pkg/context/http"
-	"github.com/prometheus/client_golang/prometheus"
-)
-
-type Scraper interface {
-	// Name of the Scraper.
-	Name() string
-
-	// Help describes the role of the Scraper.
-	Help() string
-
-	// Scrape collects data from database connection and sends it over channel as prometheus metric.
-	Scrape(ctx *contextHttp.HttpContext, ch chan<- prometheus.Metric) error
-}
